+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 configuration in RPM and DEB packages (#303)
 * Fix `lc-admin` numerical outputs such as file completion percentage showing as
 1.00e2 instead of 100 (#304)
+* Improve debug logging of backoff calculations
 
 ## 2.0.0